# openapi.yaml
openapi: 3.0.3
info:
  title: Event Attendance Management System (EAMS) API
  description: >
    API-only backend for the Event Attendance Management System. Multi-organization tenant
    model: every resource except `User` (Super Admins have no organization) is scoped to the
    caller's organization automatically — Super Admins see across all organizations, everyone
    else only ever sees their own. Cross-organization access to a scoped resource returns
    `404 Not Found` (not `403`), since the record is filtered out of the query entirely.
  version: "1.0.0"
servers:
  - url: http://localhost:8000/api/v1
    description: Local development (docker compose)
security:
  - bearerAuth: []
...
paths:
...
  /attendees/{attendee}:
    parameters:
      - $ref: "#/components/parameters/AttendeeId"
    get:
      tags: [Attendees]
      summary: Get an attendee
      responses:
        "200": { $ref: "#/components/responses/AttendeeItem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/NotFound" }
  /attendees/{attendee}/photo:
    parameters:
      - $ref: "#/components/parameters/AttendeeId"
    post:
      tags: [Attendees]
      summary: Upload an attendee's photo
      description: >
        Creates 4 WebP images (sm/md/lg/original) from the uploaded PNG/JPEG. Org Admin+ only.
        Supports both multipart file uploads and base64 input (with or without data-URI prefix).
        Decoded size must not exceed 10MB.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  format: binary
                  description: PNG or JPEG image file
          application/json:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  description: "Base64-encoded PNG/JPEG, optionally prefixed with data-URI (e.g., 'data:image/png;base64,...')"
      responses:
        "200": { $ref: "#/components/responses/AttendeeItem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [Attendees]
      summary: Remove an attendee's photo
      description: Deletes all 4 image sizes. Org Admin+ only.
      responses:
        "200": { $ref: "#/components/responses/AttendeeItem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

    put:
      tags: [Attendees]
      summary: Update an attendee
      description: Does not re-run duplicate detection.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                union_id: { type: integer, nullable: true }
                mission_id: { type: integer, nullable: true }
                church_id: { type: integer, nullable: true }
                first_name: { type: string, maxLength: 100 }
                middle_name: { type: string, maxLength: 100, nullable: true }
                last_name: { type: string, maxLength: 100 }
                mobile_no: { type: string, maxLength: 20, nullable: true }
                email_address: { type: string, format: email, nullable: true }
                remarks: { type: string, nullable: true }
      responses:
        "200": { $ref: "#/components/responses/AttendeeItem" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [Attendees]
      summary: Delete an attendee
      description: Org Admin+ only.
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------------------------------------------------------------- Events
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Laravel Sanctum personal access token, obtained from `POST /auth/login`.
  parameters:
...
    AttendeeId:
      { name: attendee, in: path, required: true, schema: { type: integer } }
...
  schemas:
...
    Union:
      type: object
      properties:
        id: { type: integer }
        organization_id: { type: integer }
        name: { type: string }
        code: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Mission:
      type: object
      properties:
        id: { type: integer }
        organization_id: { type: integer }
        union_id: { type: integer }
        name: { type: string }
        code: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Church:
      type: object
      properties:
        id: { type: integer }
        organization_id: { type: integer }
        union_id: { type: integer }
        mission_id: { type: integer }
        name: { type: string }
        address: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ImageUrls:
      type: object
      description: URLs for 4 WebP image sizes. All are null if photo/banner was never uploaded.
      properties:
        sm: { type: string, format: uri, nullable: true, description: "480px width" }
        md: { type: string, format: uri, nullable: true, description: "1024px width" }
        lg: { type: string, format: uri, nullable: true, description: "1920px width" }
        original: { type: string, format: uri, nullable: true, description: "Up to 3000px width, maintains aspect ratio" }
    Attendee:
      type: object
      properties:
        id: { type: integer }
        organization_id: { type: integer }
        union_id: { type: integer, nullable: true }
        mission_id: { type: integer, nullable: true }
        church_id: { type: integer, nullable: true }
        first_name: { type: string }
        middle_name: { type: string, nullable: true }
        last_name: { type: string }
        full_name:
          {
            type: string,
            description: "Derived: first + middle + last, trimmed.",
          }
        photo_urls: { $ref: "#/components/schemas/ImageUrls" }
        union: { $ref: "#/components/schemas/Union" }
        mission: { $ref: "#/components/schemas/Mission" }
        church: { $ref: "#/components/schemas/Church" }
        mobile_no: { type: string, maxLength: 20, nullable: true }
        email_address: { type: string, format: email, nullable: true }
        remarks: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ValidationError:
      type: object
      properties:
        message: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
    Error:
      type: object
      properties:
        message: { type: string }
  responses:
    Unauthenticated:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { message: "Unauthenticated." }
    Forbidden:
      description: Authenticated, but not allowed to perform this action
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { message: "This action is unauthorized." }
    NotFound:
      description: >
        Not found — either the record truly doesn't exist, or (for organization-scoped
        resources) it belongs to a different organization than the caller's.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
...
    AttendeeItem:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { $ref: "#/components/schemas/Attendee" }
...